+++ BE/src/auth/auth.module.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Module } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
@@ -6,14 +6,10 @@ import { UsersModule } from 'src/users/users.module';
 import { JwtModule } from '@nestjs/jwt';
 import { GoogleStrategy } from './google.strategy';
 import { MulterModule } from '@nestjs/platform-express';
-import * as path from 'path';
-import * as multerS3 from 'multer-s3';
-import { S3Client } from '@aws-sdk/client-s3';
-import { v4 as uuid } from 'uuid';
+import { multerConfig } from 'src/common/multer.config';
 
 @Module({
   imports: [
-    ConfigModule.forRoot(),
     JwtModule.registerAsync({
       imports: [ConfigModule],
       inject: [ConfigService],
@@ -22,26 +18,10 @@ import { v4 as uuid } from 'uuid';
         signOptions: { expiresIn: configService.get('JWT_EXPIRES_IN') },
       }),
     }),
-    MulterModule.register({
-      fileFilter: (req, file, callback) => {
-        const ext = path.extname(file.originalname);
-        if (ext !== '.jpg' && ext !== '.jpeg' && ext !== '.png') {
-          return callback(
-            new BadRequestException('jpg/jpeg/png 파일만 업로드 가능합니다!'),
-            false,
-          );
-        }
-        return callback(null, true);
-      },
-      limits: { fileSize: 1024 * 1024 * 10 }, //10MB
-      storage: multerS3({
-        s3: new S3Client({}),
-        bucket: process.env['S3_BUCKET_NAME'],
-        key: function (req, file, callback) {
-          const fileExtension = path.extname(file.originalname);
-          callback(null, `public/IMG_${uuid()}${fileExtension}`);
-        },
-      }),
+    MulterModule.registerAsync({
+      imports: [ConfigModule],
+      inject: [ConfigService],
+      useFactory: multerConfig,
     }),
     UsersModule,
   ],
