+++ BE/src/common/const/env-keys.const.ts
@@ -1,0 +1,7 @@
+export enum ENV {
+  IMAGE_ENDPOINT = 'IMAGE_ENDPOINT',
+  IMAGE_ACCESSKEY = 'IMAGE_ACCESSKEY',
+  IMAGE_SECRETKEY = 'IMAGE_SECRETKEY',
+  IMAGE_REGION = 'IMAGE_REGION',
+  IMAGE_BUCKET = 'IMAGE_BUCKET',
+}
